# docs/swagger-docs/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/activity": {
      "get": {
        "tags": [
          "UserActivity-Controller"
        ],
        "operationId": "getAllUserActivitiesBasedOnEventType",
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          },
          {
            "name": "arg0",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UserActivityRequest"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserActivityRequest"
              },
              "example": {
                "data": {
                  "id": "13453",
                  "type": "activity",
                  "attributes": {
                    "eventType": "sign-in"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOListUserActivity"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "UserActivity-Controller"
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          }
        ],
        "operationId": "createUserActivity",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserActivityRequest"
              },
              "example": {
                "data": {
                  "id": "13453",
                  "type": "activity",
                  "attributes": {
                    "eventType": "Sign-in",
                    "userId": "10",
                    "ipAddress": "192.168.9.5",
                    "eventDetails": "User signed in"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": [
          "User-Controller"
        ],
        "operationId": "retrieveUser",
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "User-Controller"
        ],
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {}
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/activity/{param}": {
      "get": {
        "tags": [
          "UserActivity-Controller"
        ],
        "operationId": "getUserActivityForSpecificUser",
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          },
          {
            "name": "param",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "arg1",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestDTO"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserActivityRequest"
              },
              "example": {
                "data": {
                  "id": "13453",
                  "type": "activity",
                  "attributes": {
                    "startDate": "2024-10-10",
                    "endDate": "2024-10-18",
                    "eventType": "Sign-in",
                    "email": "vishwasransingh@gmail.com",
                    "mobile": "7743898263"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/partners/banks/activity/{partnerId}": {
      "post": {
        "tags": [
          "Partner-Controller"
        ],
        "operationId": "getUserActivityForSpecificPartner",
        "parameters": [
          {
            "name": "partnerId",
            "in": "path",
            "required": true,
            "description": "ID of the partner",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartnerRequest"
              },
              "example": {
                "data": {
                  "type": "partner",
                  "attributes": {
                    "activityEventType": "Sign-in"
                  }
                }
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "description": "Activity data for a specific partner",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/partners/banks/activity/all": {
      "post": {
        "tags": [
          "Partner-Controller"
        ],
        "operationId": "getAllActivitiesBasedOnEventType",
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartnerRequest"
              },
              "example": {
                "data": {
                  "type": "partner",
                  "attributes": {
                    "activityEventType": "Sign-in"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "All activities for a given event type",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "RequestDTO": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RequestData"
          }
        }
      },
      "RequestData": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "attributes": {
            "type": "object"
          }
        }
      },
      "DataDTOObject": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "attributes": {
            "type": "object"
          }
        }
      },
      "ErrorDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          }
        }
      },
      "MetaDTOObject": {
        "type": "object",
        "properties": {
          "attributes": {
            "type": "object"
          }
        }
      },
      "ResponseDTOObject": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DataDTOObject"
          },
          "meta": {
            "$ref": "#/components/schemas/MetaDTOObject"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDTO"
            }
          }
        }
      },
      "DataDTOListUserActivity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "attributes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserActivityRequest"
            }
          }
        }
      },
      "MetaDTOListUserActivity": {
        "type": "object",
        "properties": {
          "attributes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserActivityRequest"
            }
          }
        }
      },
      "ResponseDTOListUserActivity": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DataDTOListUserActivity"
          },
          "meta": {
            "$ref": "#/components/schemas/MetaDTOListUserActivity"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDTO"
            }
          }
        }
      },
      "UserActivityRequest": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "type": {
                "type": "string"
              },
              "attributes": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string",
                    "pattern": "^[0-9]{1,6}$"
                  },
                  "eventType": {
                    "type": "string",
                    "pattern": "^[a-zA-Z]{5,50}$"
                  },
                  "eventDetails": {
                    "type": "string",
                    "pattern": "^[a-zA-Z]{5,100}$"
                  },
                  "startdate": {
                    "type": "string",
                    "format": "date"
                  },
                  "endDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "email": {
                    "type": "string",
                    "pattern": "^[A-Za-z0-9+_.-]+@(.+)$"
                  },
                  "mobile": {
                    "type": "string",
                    "pattern": "^[1-9]\\d{9}$"
                  }
                }
              }
            }
          }
        }
      },
      "RequestId": {
        "type": "string",
        "pattern": "^[a-zA-Z0-9]",
        "description": "UUID for each request"
      },
...
      "PartnerRequest": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "type": {
                "type": "string"
              },
              "attributes": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "pattern": "^[A-Za-z0-9+_.-]+@(.+)$",
                    "description": "User's email address"
                  },
                  "mobile": {
                    "type": "string",
                    "pattern": "^[1-9]\\d{9}$",
                    "description": "User's mobile number"
                  },
                  "password": {
                    "type": "string",
                    "patern": "",
                    "description": "Password"
                  },
                  "emailOtp": {
                    "type": "string",
                    "pattern": "^\\d{6}$",
                    "description": "OTP sent to the user's email, exactly 6 digits."
                  },
                  "mobileOtp": {
                    "type": "string",
                    "pattern": "^\\d{6}$",
                    "description": "OTP sent to the user's mobile number, exactly 6 digits."
                  },
                  "ipAddress": {
                    "type": "string",
                    "description": "IP address."
                  },
                  "companyName": {
                    "type": "string",
                    "pattern": "^[\\w\\s]+$",
                    "description": "Name of the company; only letters, numbers, and spaces."
                  },
                  "companyAddress": {
                    "type": "string",
                    "description": "Address of the company."
                  },
                  "dateOfIncorporation": {
                    "type": "string",
                    "format": "date"
                  },
                  "license": {
                    "type": "string",
                    "pattern": "^[A-Za-z0-9]+$",
                    "description": "License number - alphanumeric characters."
                  },
                  "cin": {
                    "type": "string",
                    "pattern": "^[A-Za-z0-9]+$"
                  },
                  "gstNo": {
                    "type": "string",
                    "pattern": "^[A-Za-z0-9]+$"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}